requirements.txt
appnope==0.1.2
asgiref==3.3.4
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.7
Django==3.2
djangorestframework==3.12.4
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==8.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
iniconfig==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
zipp==3.4.1
